attempt to fix tests for webpack3
mini-css-extract-plugin does not work AT ALL with webpack3. The old
configuration for extract-text-plugin should work, but I'm missing
something, because the webpack build never completes, and just hangs.
I've tested in an actual webpack3 project and verified it works, so
there's something in my test setup, not the project code itself. Will
eventually fix, but I don't want to hold up an actual fix to fix a
test...